Updated logic for floats

+++ src/api/payments/venmo/schemas.py
@@ -1,7 +1,7 @@
 from pydantic import BaseModel
 
 class VenmoBase(BaseModel):
-    payment_amount:int
+    payment_amount:float
     note:str
     username: list[str]
     payment_type: str
